+++ Sources/QuickLMDB/Database/Database.swift
@@ -39,7 +39,7 @@ public struct Database:Sendable, MDB_db_basic {
 	/// 	- name: the name of the database. you may pass `nil` for this argument if you plan on storing only one database in the environment.
 	/// 	- flags: the flags that will be used when opening the database.
 	///		- tx: a pointer to the transaction that will be used to open the database.
-    public init(env:borrowing Environment, name name_in:String?, flags:MDB_db_flags, tx:borrowing Transaction) throws {
+    public init(env:borrowing Environment, name name_in:String?, flags:MDB_db_flags, tx:borrowing Transaction) throws(LMDBError) {
 		var mutateLogger = env.logger()
 		mutateLogger?[metadataKey:"type"] = "Database.Strict<\(String(describing:MDB_db_key_type.self)), \(String(describing:MDB_db_val_ptrtype.self))>"
 		self._logger = mutateLogger
@@ -59,7 +59,7 @@ public struct Database:Sendable, MDB_db_basic {
 	/// 	- name: the name of the database. you may pass `nil` for this argument if you plan on storing only one database in the environment.
 	/// 	- flags: the flags that will be used when opening the database.
 	///		- tx: a pointer to the transaction that will be used to open the database.
-	public init(env:borrowing Environment, name name_in:String?, flags:MDB_db_flags, tx:borrowing Transaction) throws {
+	public init(env:borrowing Environment, name name_in:String?, flags:MDB_db_flags, tx:borrowing Transaction) throws(LMDBError) {
 		self._db_env = copy env
 		self._db_name = name_in
 		var dbHandle = MDB_dbi()
@@ -149,7 +149,7 @@ extension Database {
 		/// 	- name: the name of the database. you may pass `nil` for this argument if you plan on storing only one database in the environment.
 		/// 	- flags: the flags that will be used when opening the database.
 		///		- tx: a pointer to the transaction that will be used to open the database.
-		public init(env:borrowing Environment, name:String?, flags:consuming MDB_db_flags, tx:borrowing Transaction) throws {
+		public init(env:borrowing Environment, name:String?, flags:consuming MDB_db_flags, tx:borrowing Transaction) throws(LMDBError) {
 			flags.update(with:.dupSort)
 			var mutateLogger = env.logger()
 			mutateLogger?[metadataKey:"type"] = "Database.Strict<\(String(describing:K.self)), \(String(describing:V.self))>"
@@ -166,7 +166,7 @@ extension Database {
 			MDB_db_assign_compare_val_f(db:self, type:MDB_db_val_type.self, tx:tx)
 		}
 		#else
-		public init(env:borrowing Environment, name:String?, flags:consuming MDB_db_flags, tx:borrowing Transaction) throws {
+		public init(env:borrowing Environment, name:String?, flags:consuming MDB_db_flags, tx:borrowing Transaction) throws(LMDBError) {
 			flags.update(with:.dupSort)
 			self._db_env = copy env
 			self._db_name = name
@@ -225,7 +225,7 @@ extension Database {
 		/// 	- name: the name of the database. you may pass `nil` for this argument if you plan on storing only one database in the environment.
 		/// 	- flags: the flags that will be used when opening the database.
 		///		- tx: borrows a transaction that will be used to complete the database initialization.
-		public init(env:borrowing Environment, name:String?, flags:consuming MDB_db_flags, tx:borrowing Transaction) throws {
+		public init(env:borrowing Environment, name:String?, flags:consuming MDB_db_flags, tx:borrowing Transaction) throws(LMDBError) {
 			// configure the correct flags before consuming the variable
 			flags.update(with:.dupFixed)
 			flags.update(with:.dupSort)
@@ -246,7 +246,7 @@ extension Database {
 			MDB_db_assign_compare_val_f(db:self, type:MDB_db_val_type.self, tx:tx)
 		}
 		#else
-		public init(env:borrowing Environment, name:String?, flags:consuming MDB_db_flags, tx:borrowing Transaction) throws {
+		public init(env:borrowing Environment, name:String?, flags:consuming MDB_db_flags, tx:borrowing Transaction) throws(LMDBError) {
 			flags.update(with:.dupFixed)
 			flags.update(with:.dupSort)
 			self._db_env = copy env
@@ -304,7 +304,7 @@ extension Database {
 		/// 	- name: the name of the database. you may pass `nil` for this argument if you plan on storing only one database in the environment.
 		/// 	- flags: the flags that will be used when opening the database.
 		///		- tx: a pointer to the transaction that will be used to open the database.
-		public init(env:borrowing Environment, name:String?, flags: MDB_db_flags, tx:borrowing Transaction) throws {
+		public init(env:borrowing Environment, name:String?, flags: MDB_db_flags, tx:borrowing Transaction) throws(LMDBError) {
 			var mutateLogger = env.logger()
 			mutateLogger?[metadataKey:"type"] = "Database.Strict<\(String(describing:K.self)), \(String(describing:V.self))>"
 			self._db_env = copy env
@@ -323,7 +323,7 @@ extension Database {
 			MDB_db_assign_compare_key_f(db:self, type:MDB_db_key_type.self, tx:tx)
 		}
 		#else
-		public init(env:borrowing Environment, name:String?, flags:MDB_db_flags, tx:borrowing Transaction) throws {
+		public init(env:borrowing Environment, name:String?, flags:MDB_db_flags, tx:borrowing Transaction) throws(LMDBError) {
 			self._db_env = copy env
 			self._db_name = name
 			var handle = MDB_dbi()